move control_abbreviation and decorator to helper
